+++ src/Pagination/Pagination.css
